Pylenium2.0/pyproject.toml
[tool.poetry]
name = "pylenium2.0"
version = "0.1.0"
description = ""
authors = ["FelepeX"]
readme = "README.md"
packages = [{include = "pylenium 2"}]

[tool.poetry.dependencies]
python = "^3.12"
pyleniumio = "^1.20.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
